feat : 에셋 구매 및 조회 API 추가

+++ backend/src/main/java/com/thedebuggers/backend/dto/MyAssetResDto.java
@@ -13,6 +13,6 @@ import java.util.List;
 @Builder
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 public class MyAssetResDto {
-    List<Integer> avatarList;
-    List<Integer> roomList;
+    List<Long> avatarList;
+    List<Long> roomList;
 }
